+++ README.md
@@ -78,6 +78,6 @@
 ### 추가 리팩토링
 - [ ] `승패무`를 enum으로 변경한다.
 - [ ] 디미터 법칙 학습과 적용
-- [ ] 뷰 리팩토링
-  - [ ] 빠진 출력 수정
-  - [ ] 추가 카드 로직 변경
+- [x] 뷰 리팩토링
+  - [x] 빠진 출력 수정
+  - [x] 추가 카드 로직 변경
