+++ src/main/java/blackjack/view/InputView.java
@@ -10,9 +10,6 @@ public class InputView {
     private static final Scanner SCANNER = new Scanner(System.in);
     private static final String REQUEST_NAME_MESSAGE = "게임에 참여할 사람의 이름을 입력하세요. (쉼표 기준으로 분리)";
     private static final String MORE_DRAW_MESSAGE = "는 한장의 카드를 더 받겠습니까?(예는 y, 아니오는 n)";
-    private static final String YES_OR_NO_ERROR = "[ERROR] y 혹은 n만 입력할 수 있습니다.";
-    private static final List<String> YES_OR_NO_VALIDATION = new ArrayList<>(
-        Arrays.asList("y", "n", "Y", "N"));
 
     private InputView() {
     }
@@ -22,28 +19,8 @@ public class InputView {
         return SCANNER.nextLine();
     }
 
-    public static boolean requestMoreDraw(String name) {
+    public static String requestMoreDraw(String name) {
         System.out.println(name + MORE_DRAW_MESSAGE);
-        String input = SCANNER.nextLine();
-        try {
-            validateYorN(input);
-            return translateSignal(input);
-        } catch (IllegalArgumentException e) {
-            System.out.println(e.getMessage());
-            return requestMoreDraw(name);
-        }
-    }
-
-    private static boolean translateSignal(String input) {
-        if (input.equals("y")) {
-            return true;
-        }
-        return false;
-    }
-
-    private static void validateYorN(String input) {
-        if (!YES_OR_NO_VALIDATION.contains(input)) {
-            throw new IllegalArgumentException(YES_OR_NO_ERROR);
-        }
+        return SCANNER.nextLine();
     }
 }
